backend/requirements.txt
# COD-Odisha Backend Dependencies

# Core ML/DL Libraries
torch==2.0.1
torchvision==0.15.2
torchaudio==2.0.2
numpy>=1.24.0
scipy>=1.10.0
scikit-image>=0.21.0
opencv-python>=4.8.0

# Model Utilities
pillow>=10.0.0
imageio>=2.32.0

# Training & Evaluation
tqdm>=4.66.0
pyyaml>=6.0
tensorboard>=2.14.0
wandb>=0.15.0

# Data & Computer Vision
albumentations>=1.3.0
opencv-contrib-python>=4.8.0

# API & Deployment
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
pydantic>=2.0.0

# Optional: Performance monitoring
psutil>=5.9.0
